4.22 tehty

+++ osa4/blogilista/controllers/blogs.js
@@ -1,7 +1,7 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
-
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
@@ -10,25 +10,30 @@ blogsRouter.get('/', async (request, response) => {
 })
   
 blogsRouter.post('/', async (request, response, next) => {
-  try {
     const body = request.body
-    const user = await User.findById(body.userId)
-
     if(!body.title || !body.url) {
       return response.status(400).end()
     }
-
-      const blog = new Blog({
-        title: body.title,
-        author: body.author,
-        url: body.url,
-        likes: body.likes || 0,
-        user: user._id
-      })
+    try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (request.token === undefined || decodedToken.id === undefined) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    
+    const user = request.user
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes || 0,
+      user: user._id
+    })
 
     const blogObject = new Blog(blog)
   
     const result = await blogObject.save()
+    user.blogs = user.blogs.concat(result.id)
+    await user.save()
     response.status(201).json(result)
   }
   catch(exception){
@@ -38,8 +43,26 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (request.token === undefined || decodedToken.id === undefined) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+    const user = request.user
+    if (blog.user.toString() === user.id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+
+      const updatedBlogs = user.blogs.filter(id => id != request.params.id)
+      user.blogs = updatedBlogs
+      await user.save()
+
+      response.status(204).end()
+    }
+    else {
+      response.status(403).json({error: 'user not authorized for this operation'})
+    }
+
   } 
   catch(exception) {
     next(exception)
